+++ src/stores/index.ts
@@ -2,15 +2,17 @@ import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
 
 import { AuthSlice, createAuthSlice } from './authSlice';
+import { createHeaderSlice, HeaderSlice } from './headerSlice';
 import { UserSlice, createUserSlice } from './userSlice';
 
-type StoreState = AuthSlice & UserSlice;
+type StoreState = AuthSlice & UserSlice & HeaderSlice;
 
 const useBoundStore = create<StoreState>()(
   persist(
     (set, get, api) => ({
       ...createAuthSlice(set, get, api),
       ...createUserSlice(set, get, api),
+      ...createHeaderSlice(set, get, api),
     }),
     {
       name: 'abcdedu',
